Fixes #24214 - fill in vars on repo URLs
Previously, repo baseurls were sent as-is from redhat.repo to
Satellite. This resulted in the client reporting it was subscribed to
repos that don't exist, since the URLs would have the literal string
`$releasever/$basearch` in it.

This commit adds dnf and yum code to populate `$releasever` and
`$basearch`. It will attempt to use DNF libraries first, and if that
raises an `ImportError`, it will try yum.