Fixes #27650 - Handles combined profile uploads
Subscription Manager 1.23-2 and above have the ability to generate a
profile.json that combines package profile, enabled repositories profile
and modulemd profiles. Katello Host tools at present is unable to take
advantage of this facility. So 2 sets of reports related to enabled
repositories are sent to the server.
This commit does the following
1) Removes the package_profile_on_trans check and instead always defers
to subscription manager to handle it
2) Checks if combined profiles feature is enabled by checking if
rhsm.profile.EnabledRepos is available.
3) If Combined profiles are available then defers the enabled repos
upload to subscription manager as well
4) Makes the remove_cache rm the profile.json instead of whereever the
cache was before. This remove_cache is used in the "force" calls.
katello-package-upload -f and katello-enabled-repos-upload -f
operations.

Subscription manager sends information to the server only if the cache
changes and hence unnecessary calls to the server are avoided.