Rewrite scripts to use console_scripts entry points (#125)
* Drop EL5

The Python version in EL5 can no longer communicate with PyPI since it
doesn't support modern TLS. RHLE 5 also went EOL some time ago so it's
time to stop caring about this.

* Order images to newest first

This way the newest OS gets tested first, which is more likely to work
and be faster.

* Use dnf on Fedora

* Update EL6 to use Foreman 2.5 client tooling

While 2.5 is already end of life, this was the last release to support
it. That's better than 1.21.

* Update EL7 to use nightly client tools

Also pulls the image from quay and makes it a single transaction for
Foreman and EPEL.

* Update EL8 image to stream and nightly

* Only run flake8 on Python 3

flake8 4.0.0 requires Python >= 3.6 and this was the easiest way to keep
it working.

* Clean up metadata

This adds an URL which and removes the empty requires variable in favor
an inline definition.

* Rewrite scripts to use console_scripts entry points

This way Python makes the appropriate wrapper itself. This means it can
be uploaded to PyPI and consumed that way.